tests/wpt/tests/css/css-position/parsing/inset-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Positioned Layout Module: parsing inset with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-position/#insets">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
  const single_values = [
    "0deg",
    "calc(20deg)",
    "10"
  ];
  const double_values = [
    "inherit auto",
    "inherit inherit"
  ];
  function test_inset_longhand(longhand_property) {
    for (let value of single_values)
      test_invalid_value(longhand_property, value);
  }
  function test_inset_shorthand(shorthand_property) {
    for (let value of single_values)
      test_invalid_value(shorthand_property, value);
    for (let value of double_values)
      test_invalid_value(shorthand_property, value);
  }

  test_inset_longhand("inset-block-start");
  test_inset_longhand("inset-block-end");
  test_inset_longhand("inset-inline-start");
  test_inset_longhand("inset-inline-end");

  test_inset_shorthand("inset-block");
  test_inset_shorthand("inset-inline");
</script>
</body>
</html>
